Feat: Integracion del Middleware

+++ routes/web.php
@@ -19,19 +19,23 @@ Route::get('/', function () {
 
 Auth::routes();
 
+Route::middleware(['auth', 'admin'])->group(function () {
+    Route::get('/especialidades', [App\Http\Controllers\admin\SpecialtyController::class, 'index']);
+
+    Route::get('/especialidades/create', [App\Http\Controllers\admin\SpecialtyController::class, 'create']);
+    Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\admin\SpecialtyController::class, 'edit']);
+    Route::post('/especialidades', [App\Http\Controllers\admin\SpecialtyController::class, 'sendData']);
+
+    Route::put('/especialidades/{specialty}', [App\Http\Controllers\admin\SpecialtyController::class, 'update']);
+    Route::delete('/especialidades/{specialty}', [App\Http\Controllers\admin\SpecialtyController::class, 'destroy']);
+
+    //Rutas de Medicos
+    Route::resource('medicos','App\Http\Controllers\admin\DoctorController');
+
+    //Rutas de Pacientes
+    Route::resource('pacientes','App\Http\Controllers\admin\PatientController');
+});
+
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-Route::get('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'index']);
 
-Route::get('/especialidades/create', [App\Http\Controllers\SpecialtyController::class, 'create']);
-Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\SpecialtyController::class, 'edit']);
-Route::post('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'sendData']);
-
-Route::put('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'update']);
-Route::delete('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'destroy']);
-
-//Rutas de Medicos
-Route::resource('medicos','App\Http\Controllers\DoctorController');
-
-//Rutas de Pacientes
-Route::resource('pacientes','App\Http\Controllers\PatientController');
